Fixes to enable building with BR_EMBEDDED
