+++ openbr/core/utility.cpp
@@ -1,0 +1,34 @@
+#include <openbr/core/qtutils.h>
+#include "utility.h"
+
+QStringList br::getFiles(QDir dir, bool recursive)
+{
+    dir = QDir(dir.canonicalPath());
+
+    QStringList files;
+    foreach (const QString &file, QtUtils::naturalSort(dir.entryList(QDir::Files)))
+        files.append(dir.absoluteFilePath(file));
+
+    if (!recursive) return files;
+
+    foreach (const QString &folder, QtUtils::naturalSort(dir.entryList(QDir::Dirs | QDir::NoDotAndDotDot))) {
+        QDir subdir(dir);
+        bool success = subdir.cd(folder); if (!success) qFatal("cd failure.");
+        files.append(getFiles(subdir, true));
+    }
+    return files;
+}
+
+QStringList br::getFiles(const QString &regexp)
+{
+    QFileInfo fileInfo(regexp);
+    QDir dir(fileInfo.dir());
+    QRegExp re(fileInfo.fileName());
+    re.setPatternSyntax(QRegExp::Wildcard);
+
+    QStringList files;
+    foreach (const QString &fileName, dir.entryList(QDir::Files))
+        if (re.exactMatch(fileName))
+            files.append(dir.filePath(fileName));
+    return files;
+}
