+++ openbr/gui/utility.cpp
@@ -3,7 +3,6 @@
 #include <assert.h>
 #include <opencv2/imgproc/imgproc.hpp>
 #include <opencv2/imgproc/imgproc_c.h>
-#include <openbr/core/qtutils.h>
 #include "utility.h"
 
 using namespace cv;
@@ -45,36 +44,4 @@ QImage br::toQImage(const Mat &mat)
     else if (mat8u.channels() == 1) cvtColor(mat8u, mat8uc3, CV_GRAY2RGB);
 
     return QImage(mat8uc3.data, mat8uc3.cols, mat8uc3.rows, 3*mat8uc3.cols, QImage::Format_RGB888).copy();
-}
-
-QStringList br::getFiles(QDir dir, bool recursive)
-{
-    dir = QDir(dir.canonicalPath());
-
-    QStringList files;
-    foreach (const QString &file, QtUtils::naturalSort(dir.entryList(QDir::Files)))
-        files.append(dir.absoluteFilePath(file));
-
-    if (!recursive) return files;
-
-    foreach (const QString &folder, QtUtils::naturalSort(dir.entryList(QDir::Dirs | QDir::NoDotAndDotDot))) {
-        QDir subdir(dir);
-        bool success = subdir.cd(folder); if (!success) qFatal("cd failure.");
-        files.append(getFiles(subdir, true));
-    }
-    return files;
-}
-
-QStringList br::getFiles(const QString &regexp)
-{
-    QFileInfo fileInfo(regexp);
-    QDir dir(fileInfo.dir());
-    QRegExp re(fileInfo.fileName());
-    re.setPatternSyntax(QRegExp::Wildcard);
-
-    QStringList files;
-    foreach (const QString &fileName, dir.entryList(QDir::Files))
-        if (re.exactMatch(fileName))
-            files.append(dir.filePath(fileName));
-    return files;
 }
