+++ openbr/gui/utility.h
@@ -1,5 +1,5 @@
-#ifndef BR_UTILITY_H
-#define BR_UTILITY_H
+#ifndef BR_GUI_UTILITY_H
+#define BR_GUI_UTILITY_H
 
 #include <QImage>
 #include <QStringList>
@@ -12,9 +12,7 @@ namespace br
 {
 
 BR_EXPORT QImage toQImage(const cv::Mat &mat);
-BR_EXPORT QStringList getFiles(QDir dir, bool recursive);
-BR_EXPORT QStringList getFiles(const QString &regexp);
 
 } // namespace br
 
-#endif // BR_UTILITY_H
+#endif // BR_GUI_UTILITY_H
